+++ webapp/app/components/error-box/component.js
@@ -2,8 +2,10 @@ import Ember from "ember";
 
 export default Ember.Component.extend({
   expanded: false,
+  error: null,
 
-  classNameBindings: ['expanded'],
+  classNames: ['error-box'],
+  classNameBindings: ['expanded', 'error.is_interruption:interruption'],
 
   actions: {
     toggle_expanded() {