+++ webapp/app/session/test/interruptions/controller.js
@@ -1,0 +1,5 @@
+import PaginatedFilteredController
+  from "../../../controllers/paginated_filtered_controller";
+import RelativeItemJump from "../../../mixins/relative-item-jump";
+
+export default PaginatedFilteredController.extend(RelativeItemJump, {});